Revert "PHR-8766 Add interfaces for all services in testsupport module. "
