+++ testsupport/src/main/java/com/pkb/common/testsupport/config/ITestControlServiceConfig.java
@@ -1,6 +1,10 @@
 package com.pkb.common.testsupport.config;
 
+import java.util.Optional;
+import java.util.Set;
+
 import com.pkb.common.ClearableInternalState;
+import com.pkb.common.config.BaseConfig;
 import com.pkb.common.config.ConfigStorage;
 import com.pkb.common.datetime.DateTimeService;
 import com.pkb.common.testlogging.DetailLoggingProvider;
@@ -11,9 +15,6 @@ import com.pkb.common.testsupport.services.MoveTimeService;
 import com.pkb.common.testsupport.services.PubSubNamespaceService;
 import com.pkb.common.testsupport.services.SetFixedTimestampService;
 import com.pkb.common.testsupport.services.ToggleDetailedLoggingService;
-
-import java.util.Optional;
-import java.util.Set;
 
 public interface ITestControlServiceConfig {
     Optional<String> getEmulatorEndpoint();
@@ -39,8 +40,10 @@ public interface ITestControlServiceConfig {
     boolean getShouldStartListener();
 
     DateTimeService getDateTimeService();
-    
+
     ConfigStorage getConfigStorage();
+
+    BaseConfig getBaseConfig();
 
     Set<ClearableInternalState> getClearables();
