--- testsupport/src/main/java/com/pkb/common/testsupport/services/DefaultToggleDetailedLoggingService.java
@@ -1,24 +1,0 @@
-package com.pkb.common.testsupport.services;
-
-import com.pkb.common.testlogging.DetailLoggingProvider;
-import com.pkb.pubsub.testsupport.payload.ToggleDetailedLoggingRequest;
-import com.pkb.pubsub.testsupport.payload.ToggleDetailedLoggingResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class DefaultToggleDetailedLoggingService implements ToggleDetailedLoggingService {
-    private static final Logger LOGGER = LoggerFactory.getLogger(java.lang.invoke.MethodHandles.lookup().lookupClass());
-    private final DetailLoggingProvider testLoggingService;
-
-    public DefaultToggleDetailedLoggingService(DetailLoggingProvider testLoggingService) {
-        this.testLoggingService = testLoggingService;
-    }
-
-    @Override
-    public ToggleDetailedLoggingResponse process(ToggleDetailedLoggingRequest message) {
-        LOGGER.info("ToggleDetailedLoggingService.process message received");
-        testLoggingService.setDetailedLoggingRequired(message.getEnableDetailedLogging());
-        LOGGER.info("ToggleDetailedLoggingService.process done.");
-        return ToggleDetailedLoggingResponse.newBuilder().setSuccess(true).build();
-    }
-}
